stop tracking target
