+++ pom.xml
@@ -3,7 +3,8 @@
   <modelVersion>4.0.0</modelVersion>
   <groupId>ru.eternallyu</groupId>
   <artifactId>weather-app</artifactId>
-  <packaging>war</packaging>
+  <packaging>wa
+    r</packaging>
   <version>1.0-SNAPSHOT</version>
   <name>weather-app Maven Webapp</name>
   <url>http://maven.apache.org</url>
